+++ log4j/converter-log4j2/src/main/java/de/huxhorn/lilith/log4j2/producer/Log4j2LoggingConverter.java
@@ -1,9 +1,44 @@
+/*
+ * Lilith - a log event viewer.
+ * Copyright (C) 2007-2015 Joern Huxhorn
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
+ * Copyright 2007-2015 Joern Huxhorn
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package de.huxhorn.lilith.log4j2.producer;
 
 import de.huxhorn.lilith.data.converter.Converter;
 import de.huxhorn.lilith.data.eventsource.LoggerContext;
 import de.huxhorn.lilith.data.logging.ExtendedStackTraceElement;
 import de.huxhorn.lilith.data.logging.LoggingEvent;
+import de.huxhorn.lilith.data.logging.Marker;
 import de.huxhorn.lilith.data.logging.Message;
 import de.huxhorn.lilith.data.logging.ThreadInfo;
 import de.huxhorn.lilith.data.logging.ThrowableInfo;
@@ -153,7 +188,44 @@ public class Log4j2LoggingConverter
 			result.setThrowable(convert(log4jEvent.getThrownProxy()));
 		}
 
+		initMarker(log4jEvent, result);
+
 		return result;
+	}
+
+	private void initMarker(org.apache.logging.log4j.core.LogEvent src, LoggingEvent dst)
+	{
+		org.apache.logging.log4j.Marker origMarker = src.getMarker();
+		if(origMarker == null)
+		{
+			return;
+		}
+		Map<String, Marker> markers = new HashMap<>();
+		dst.setMarker(initMarkerRecursive(origMarker, markers));
+	}
+
+	private Marker initMarkerRecursive(org.apache.logging.log4j.Marker origMarker, Map<String, Marker> markers)
+	{
+		if(origMarker == null)
+		{
+			return null;
+		}
+		String name = origMarker.getName();
+		if(markers.containsKey(name))
+		{
+			return markers.get(name);
+		}
+		Marker newMarker = new Marker(name);
+		markers.put(name, newMarker);
+		org.apache.logging.log4j.Marker[] parents = origMarker.getParents();
+		if(parents != null)
+		{
+			for (org.apache.logging.log4j.Marker current : parents)
+			{
+				newMarker.add(initMarkerRecursive(current, markers));
+			}
+		}
+		return newMarker;
 	}
 
 	private ThrowableInfo convert(org.apache.logging.log4j.core.impl.ThrowableProxy thrown)